+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Compiler/AddSecurityVotersPassTest.php
@@ -14,6 +14,7 @@ namespace Symfony\Bundle\SecurityBundle\Tests\DependencyInjection\Compiler;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bundle\SecurityBundle\DependencyInjection\Compiler\AddSecurityVotersPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManager;
 use Symfony\Component\Security\Core\Authorization\Voter\Voter;
@@ -67,6 +68,73 @@ class AddSecurityVotersPassTest extends TestCase
         $this->assertEquals(new Reference('highest_prio_service'), $refs[0]);
         $this->assertEquals(new Reference('lowest_prio_service'), $refs[1]);
         $this->assertCount(4, $refs);
+    }
+
+    /**
+     * Test that in debug mode, voters are correctly decorated.
+     */
+    public function testThatVotersAreDecoratedInDebugMode(): void
+    {
+        $container = new ContainerBuilder();
+
+        $voterDef1 = new Definition(Voter::class);
+        $voterDef1->addTag('security.voter');
+        $container->setDefinition('voter1', $voterDef1);
+
+        $voterDef2 = new Definition(Voter::class);
+        $voterDef2->addTag('security.voter');
+        $container->setDefinition('voter2', $voterDef2);
+
+        $container
+            ->register('security.access.decision_manager', AccessDecisionManager::class)
+            ->addArgument(array($voterDef1, $voterDef2));
+        $container->setParameter('kernel.debug', true);
+
+        $compilerPass = new AddSecurityVotersPass();
+        $compilerPass->process($container);
+
+        $this->assertTrue($container->has('security.voter.debug.voter1'), 'voter1: missing decorated voter');
+
+        $def1 = $container->getDefinition('security.voter.debug.voter1');
+        $this->assertEquals(array('voter1', null, 0), $def1->getDecoratedService(), 'voter1: wrong return from getDecoratedService');
+        $this->assertEquals(new Reference('security.voter.debug.voter1.inner'), $def1->getArgument(0), 'voter1: wrong decorator argument');
+        $this->assertFalse($def1->isPublic(), 'voter1:: wrong decorator visibility');
+
+        $this->assertTrue($container->has('security.voter.debug.voter2'), 'voter2: missing decorated voter');
+
+        $def2 = $container->getDefinition('security.voter.debug.voter2');
+        $this->assertEquals(array('voter2', null, 0), $def2->getDecoratedService(), 'voter2: wrong return from getDecoratedService');
+        $this->assertEquals(new Reference('security.voter.debug.voter2.inner'), $def2->getArgument(0), 'voter2: wrong decorator argument');
+        $this->assertFalse($def2->isPublic(), 'voter2: wrong decorator visibility');
+
+        $voters = $container->findTaggedServiceIds('security.voter');
+        $this->assertCount(2, $voters, 'Incorrect count of voters');
+    }
+
+    /**
+     * Test that voters are not decorated if the application is not in debug mode.
+     */
+    public function testThatVotersAreNotDecoratedWithoutDebugMode(): void
+    {
+        $container = new ContainerBuilder();
+
+        $voterDef1 = new Definition(Voter::class);
+        $voterDef1->addTag('security.voter');
+        $container->setDefinition('voter1', $voterDef1);
+
+        $voterDef2 = new Definition(Voter::class);
+        $voterDef2->addTag('security.voter');
+        $container->setDefinition('voter2', $voterDef2);
+
+        $container
+            ->register('security.access.decision_manager', AccessDecisionManager::class)
+            ->addArgument(array($voterDef1, $voterDef2));
+
+        $compilerPass = new AddSecurityVotersPass();
+        $compilerPass->process($container);
+
+        $this->assertFalse($container->has('security.voter.debug.voter1'), 'voter1 should not be decorated');
+        $this->assertFalse($container->has('security.voter.debug.voter2'), 'voter2 should not be decorated');
     }
 
     /**